PGO-related fixes
Extensions where not properly compiled with PGO because the CFLAGS were
not read from the environment in "setup.py". Once that was fixed, gcc
failed with an internal compiler error (ICE) because the same source file was
used in 3 different extensions so it was compiled 3 different times in
the profile generation stage: https://gcc.gnu.org/bugzilla/show_bug.cgi?id=85759

After introducing the concept of mtime-conditional recompilation to
distutils, the ICE was avoided.

While building the software in a sandbox, on Gentoo, it became obvious
that the profiling task was accessing system-wide paths it had no
business accessing. This was also fixed.

A couple of regression tests fail randomly when the suite is run in
parallel, so we work around that by excluding them from the profiling
task.

EXTRATESTOPTS is now used inside PROFILE_TASK, to allow for custom
arguments (like "-jN" to run N jobs in parallel).

We no longer let the profiling task fail silently.