Lib/site.py: Readline: No exception if no init files, catch all OSError
In this case, don't let an exception propagate, which causes skipping the rest
of the code although it doesn't depend at all on the success of this operation.

Happens on systems where not default 'inputrc' is provided, such as FreeBSD
currently.

If there is no default init file (or read_init_file() was already successful,
and the last passed file was removed), don't report anything. For other OSError
exceptions (e.g., permissions), issue a warning. Don't catch other non-related
exceptions.