PyUnicode_FromFormat*: Support long long, fix possible buffer overflows
Long long can be applied with the prefix 'll' to 'd' and 'u' (these were
already the only specifiers for which 'l' was allowed, and this was not
changed).

The buffer overflows fixed are those that can be caused by very large width or
precision optional directives ('fmt' could be overflowed). To this end,
'makefmt' was modified to use snprintf and strncpy and check for any possible
overflows.

Also extended 'buffer', and replaced an hardcoded integer deduced from its size
with a calculation involving sizeof(buffer).